requirements.txt
robotframework==6.0.2
pika~=1.3.1
click~=8.1.3
robotframework-browser==16.0.2